+++ src/Terrabuild.Tests/Core/DotnetRemoteCache.fs
@@ -364,6 +364,7 @@ let private createOptions workspace variables =
       ConfigOptions.Options.Environment = None
       ConfigOptions.Options.LogTypes = []
       ConfigOptions.Options.Note = None
+      ConfigOptions.Options.GroupId = None
       ConfigOptions.Options.Label = None
       ConfigOptions.Options.Types = None
       ConfigOptions.Options.Labels = None