+++ src/Terrabuild.Tests/Core/Runner.fs
@@ -44,6 +44,7 @@ let private baseOptions workspace =
       ConfigOptions.Options.Environment = None
       ConfigOptions.Options.LogTypes = []
       ConfigOptions.Options.Note = None
+      ConfigOptions.Options.GroupId = None
       ConfigOptions.Options.Label = None
       ConfigOptions.Options.Types = None
       ConfigOptions.Options.Labels = None